AC-814 fixing 'no response' bug

+++ src/main/java/com/gft/lr/restcheck/RESTSpecLRValidator.java
@@ -1,7 +1,6 @@
 package com.gft.lr.restcheck;
 
 
-import com.gft.lr.restcheck.ifc.CommandExecutor;
 import com.gft.lr.restcheck.ifc.RESTClient;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.httpclient.HttpMethod;
@@ -16,12 +15,14 @@ import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import static java.lang.String.join;
+
 /**
- * @author: Sławomir Węgrzyn
- * @date: 17/10/2016
+ * Created on 17/10/2016
  */
 public class RESTSpecLRValidator {
 
@@ -34,20 +35,21 @@ public class RESTSpecLRValidator {
     public static final int HTTP_OK = 200;
     public static final int SEARCH_DEPTH_IS_2 = 2;
     public static final Character URL_SEPARATOR = '/';
+    public static final String WRONG_SERVER_RESPONSE = "Wrong server response: ";
 
-    private CommandExecutor commandIssuer;
+    private CommandExecutor commandExecutor;
     private String filterUrl;
     private SwaggerBuilder swaggerBuilder;
     private RESTClient restClient;
 
-    public RESTSpecLRValidator(CommandExecutor commandIssuer, RESTClient restClient, SwaggerBuilder swaggerBuilder) {
-        this.commandIssuer = commandIssuer;
+    public RESTSpecLRValidator(CommandExecutor commandExecutor, RESTClient restClient, SwaggerBuilder swaggerBuilder) {
+        this.commandExecutor = commandExecutor;
         this.restClient = restClient;
         this.swaggerBuilder = swaggerBuilder;
     }
 
-    public RESTSpecLRValidator(CommandExecutor commandIssuer, RESTClient restClient, String filterUrl, SwaggerBuilder swaggerBuilder) {
-        this(commandIssuer, restClient, swaggerBuilder);
+    public RESTSpecLRValidator(CommandExecutor commandExecutor, RESTClient restClient, String filterUrl, SwaggerBuilder swaggerBuilder) {
+        this(commandExecutor, restClient, swaggerBuilder);
         this.filterUrl = filterUrl;
     }
 
@@ -55,27 +57,46 @@ public class RESTSpecLRValidator {
         final Collection<SwaggerResource> swaggerResources = prepareJSons(filterSwaggers(prepareSwaggers()));
         final Collection<SwaggerResource> problematicJSons = new ArrayList<>();
 
-        swaggerResources.stream().forEach(swaggerResource -> {
-            File temporaryJson = null;
-            try {
-                temporaryJson = storeTempFile(prepareTempDirectory(swaggerResource), swaggerResource);
-                String sourceFilePath = getSourceFilePath(swaggerResource);
-                if (!isBackwardCompatible(temporaryJson.getAbsolutePath(), sourceFilePath)) {
-                    problematicJSons.add(swaggerResource);
-                }
-            } catch (IOException e) {
-                log.error(e.getMessage(), e);
-            } finally {
-                if (temporaryJson != null) {
-                    temporaryJson.delete();
-                    new File(temporaryJson.getParent()).delete();
-                }
-            }
-        });
+        swaggerResources.stream().filter(SwaggerResource::valid).forEach(swaggerResource -> validate(problematicJSons, swaggerResource));
+        problematicJSons.addAll(swaggerResources.stream().filter((swaggerResource) -> !swaggerResource.valid()).collect(Collectors.toList()));
 
         if (CollectionUtils.isNotEmpty(problematicJSons)) {
             throw new RESTsNotCompatibleException(problematicJSons);
         }
+    }
+
+    private void validate(Collection<SwaggerResource> problematicJSons, SwaggerResource swaggerResource) {
+        File temporaryJson = null;
+        File temporaryYaml = null;
+        try {
+            temporaryJson = storeTempFile(prepareTempDirectory(swaggerResource), swaggerResource);
+            temporaryYaml = convert2Yaml(temporaryJson);
+            String sourceFilePath = getSourceFilePath(swaggerResource);
+            StringBuffer errors = new StringBuffer();
+            if (!isBackwardCompatible(temporaryYaml.getAbsolutePath(), sourceFilePath, errors)) {
+                problematicJSons.add(new SwaggerResource(swaggerResource, errors.toString()));
+            }
+        } catch (IOException e) {
+            log.error(e.getMessage(), e);
+        } finally {
+            closeIfNotNull(temporaryJson, false);
+            closeIfNotNull(temporaryYaml, true);
+        }
+    }
+
+    private void closeIfNotNull(File temporaryJson, boolean withDirectory) {
+        if (temporaryJson != null) {
+            temporaryJson.delete();
+            if(withDirectory) {
+                new File(temporaryJson.getParent()).delete();
+            }
+        }
+    }
+
+    private File convert2Yaml(File temporaryJson) throws IOException {
+        File destYamlDir = temporaryJson.getParentFile();
+
+        return commandExecutor.convert(temporaryJson.getAbsolutePath(), destYamlDir.getAbsolutePath());
     }
 
     private Collection<SwaggerResource> filterSwaggers(Collection<SwaggerResource> swaggerResources) {
@@ -90,22 +111,36 @@ public class RESTSpecLRValidator {
         return filterUrl;
     }
 
-    private boolean isBackwardCompatible(final String temporaryJson, final String sourceFilePath) throws IOException {
-        Process process = commandIssuer.exec(temporaryJson, sourceFilePath);
+    private boolean isBackwardCompatible(final String temporaryJson,
+                                         final String sourceFilePath,
+                                         final StringBuffer outBuffer) throws IOException {
+        Process process = commandExecutor.compare(temporaryJson, sourceFilePath);
         try (BufferedReader stdInput = new BufferedReader(new InputStreamReader(process.getInputStream()));
              BufferedReader errInput = new BufferedReader(new InputStreamReader(process.getErrorStream()))) {
-            String message = getStringFromProcess(stdInput) + "\n" + getStringFromProcess(errInput);
+            process.waitFor();
+            String errors = getStringFromStream(errInput);
+            String message = join("\n", getStringFromStream(stdInput), errors);
+            outBuffer.append(message);
             log.info("Comparing tool output:\n" + message);
 
-            return StringUtils.isBlank(message);
+            return process.exitValue() == 0;
+        } catch (InterruptedException e) {
+            log.error(e.getMessage(), e);
+            throw new IOException(e);
         }
     }
 
-    private String getStringFromProcess(BufferedReader stdInput) throws IOException {
-        char[] output = new char[1024];
-        int count = stdInput.read(output);
+    private String getStringFromStream(BufferedReader stdInput) throws IOException {
+        StringBuffer sb = new StringBuffer();
+        int count;
+        char[] output = new char[128];
+        count = stdInput.read(output);
+        while (count>0) {
+            sb.append(output, 0, count);
+            count = stdInput.read(output);
+        };
 
-        return count <= 0 ? "" : new String(output);
+        return sb.toString();
     }
 
     private String getSourceFilePath(SwaggerResource swaggerResource) throws FileNotFoundException {
@@ -155,7 +190,7 @@ public class RESTSpecLRValidator {
             try {
                 int result = restClient.executeMethod(getJson);
                 if (result != HTTP_OK) {
-                    throw new IllegalStateException("Wrong server response: " + result + " for " + swaggerResource.getUrl());
+                    return new SwaggerResource(swaggerResource, WRONG_SERVER_RESPONSE + result + " for " + swaggerResource.getUrl() + ". See app logs for details.");
                 }
                 return new SwaggerResource(getJson.getResponseBodyAsString(), swaggerResource);
             } catch (IOException e) {
